build 64 bit in.rlogind and in.rshd

+++ usr/src/cmd/cmd-inet/common/kcmd.c
@@ -751,7 +751,7 @@ deswrite_compat(int fd, char *buf, int len, int secondary)
 
 	if (krb5_c_encrypt_length(kcmd_context, final_enctype,
 			use_ivecs ? (size_t)len + 4 : (size_t)len, &ret)) {
-		desoutbuf.length = ((size_t)-1);
+		desoutbuf.length = -1;
 		goto err;
 	} else {
 		desoutbuf.length = ret;
